requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Authlib==1.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
dataclasses-json==0.5.7
decorator==5.1.1
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
langchain==0.0.79
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.2
openai==0.26.5
packaging==23.0
pdfminer.six==20221105
pycparser==2.21
pydantic==1.10.5
python-dotenv==0.21.1
PyYAML==6.0
requests==2.28.2
slack-bolt==1.16.2
slack-sdk==3.20.0
SQLAlchemy==1.4.46
tenacity==8.2.1
tqdm==4.64.1
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.14
validators==0.19.0
weaviate-client==3.11.0
yarl==1.8.2
